TopComments/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
Pillow==8.1.0
praw==7.1.0
prawcore==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
wordcloud==1.8.1
